Refactor template handling in service
